requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-hub>=0.1.14

# Document processing
pypdf>=3.17.0
unstructured>=0.10.0
python-magic-bin>=0.4.14

# Vector store and embeddings
chromadb>=0.4.18
openai>=1.0.0

# Web and API tools
requests>=2.31.0
google-search-results>=2.4.2

# Cloud storage (optional - for Firebase Firestore)
google-cloud-firestore>=2.11.0
langchain-google-firestore>=0.1.0

# Optional but recommended
pandas>=2.0.0
numpy>=1.24.0